Strategy Strategy is a behavioral design pattern. It's used when we have multiple algorithms for a specific task and client decides which actual implementation should be used at runtime.
The original object, called context, holds a reference to a strategy object and delegates to it the execution of the behavior. In order to change the way the context performs its work, other objects may replace the currently linked strategy object with another one.

Advantages
You can swap algorithms used inside an object at runtime.
You can isolate the implementation details of an algorithm from the code that uses it.
You are making the system more flexible and easy to expand.
Disadvantages
If you only have a couple of algorithms, and they rarely change, there’s no real reason to come along with this pattern.
Clients must be aware of the differences between strategies to be able to select a proper one.
It increases the number of objects in the application.