Builder What is builder The builder creational pattern, is an alternative way to construct complex objects. It should be used when you want to build different objects using same object building process. Builder pattern also helps minimizing the number of parameters in constructor and thus there is no need to pass in null for optional parameters to the constructor. Certainly, the number of lines of code increase at least to double in builder pattern, but the effort pays off in terms of design flexibility and much more readable code. The parameters to the constructor are reduced and are provided in highly readable method calls.
Let’s see how we can implement builder design pattern in java:

First of all you need to create a static nested class and then copy all the arguments from the outer class to the Builder class.
Builder class should have a public constructor. It can be empty or with all the required attributes as parameters.
Builder class should have methods to set the optional parameters and it should return the same Builder object after setting the optional attribute.
The final step is to provide a build() method in the builder class that will return the Object needed by client program. For this we need to have a private constructor in the outer class with Builder class as argument.