Recursion In Java, a method that calls itself is known as a recursive method. And, this process is known as recursion.
Example
In the below example, we have called the recurse() method from inside the main method (normal method call). And, inside the recurse() method, we are again calling the same recurse method. This is a recursive call. To stop the recursive call, we need to provide some conditions inside the method. Let's see some basic example, we will sum numbers from 0 to 10 recursively:

public class MyClass {
  public static void main(String[] args) {
    int result = sum(10);
    System.out.println(result);
  }

  public static int sum(int topLimit) {
    return topLimit > 0
            ? topLimit + sum(topLimit - 1)
            : 0;
  }
}

Output: 55

Be careful, cause if the base case is not reached or not defined, then the stack overflow problem may arise:

class RecursionPractice {
    public static void main(String[] args) {
        infiniteRecursion();
    }

    private static void infiniteRecursion(){
        System.out.println("hello");
        infiniteRecursion();
    }
}

Execution result:
hello
hello
...
java.lang.StackOverflowError

Advantages and Disadvantages of Recursion
When a recursive call is made, new storage locations for variables are allocated on the stack. As each recursive call returns, the old variables and parameters are removed from the stack. Hence, recursion generally uses more memory and is generally slow. On the other hand, a recursive solution is much simpler and takes less time to write, debug and maintain.

Use cases
Common use cases for recursion are: Fibonacci sequence, factorial, tree traversal, etc.